Moving Pauli to quantum_info
