+++ qiskit/quantum_info/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2017, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+from .operators.pauli import Pauli, pauli_group
+from .states._states import basis_state, random_state, projector
+from .states._measures import state_fidelity
