+++ qiskit/quantum_info/operators/pauli.py
