+++ test/python/aer/test_extensions_simulator.py
@@ -13,7 +13,7 @@ import unittest
 import qiskit
 import qiskit.extensions.simulator
 from qiskit import Aer
-from qiskit.quantum_info.states import state_fidelity
+from qiskit.quantum_info import state_fidelity
 from qiskit import execute
 from ..common import QiskitTestCase, requires_cpp_simulator
