+++ test/python/quantum_info/test_states.py
@@ -13,9 +13,9 @@ import unittest
 import numpy as np
 from qiskit import execute, QuantumRegister, QuantumCircuit, Aer
 
-from qiskit.quantum_info.states import basis_state, random_state
-from qiskit.quantum_info.states import state_fidelity
-from qiskit.quantum_info.states import projector
+from qiskit.quantum_info import basis_state, random_state
+from qiskit.quantum_info import state_fidelity
+from qiskit.quantum_info import projector
 
 from ..common import QiskitTestCase
