+++ test/python/test_initializer.py
@@ -19,7 +19,7 @@ from qiskit import QuantumCircuit
 from qiskit import QuantumRegister
 from qiskit import ClassicalRegister
 from qiskit import execute, Aer
-from qiskit.quantum_info.states import state_fidelity
+from qiskit.quantum_info import state_fidelity
 from .common import QiskitTestCase
